refactor(survey): 설문 페이지 전용 파일 위치 정리
