+++ src/pages/survey/hooks/useSurveyViewportState.ts
